Make it possible to specify a callable to produce extra ID Token claims.
To make the extra claims more flexible, e.g., they depend on either the
user id or the client id, a callable can be specified instead of a
dict-like object. The callable must accept the user id and client id
as parameters, and produce a dict-like object with any extra claims.